Refactored from fetch in context,SingleMovie to separate useFetch file
