+++ 21-movie-db/setup/src/SingleMovie.js
@@ -1,34 +1,11 @@
 import React, { useState, useEffect } from "react";
 import { useParams, Link } from "react-router-dom";
-import { API_ENDPOINT, useGlobalContext } from "./context";
+import { useGlobalContext } from "./context";
+import useFetch from "./useFetch";
 
 const SingleMovie = () => {
-  const [movie, setMovie] = useState({});
-  const { error, loading, setError, setLoading } = useGlobalContext();
   const { id } = useParams();
-
-  const fetchMovie = async (url) => {
-    setLoading(true);
-    try {
-      const response = await fetch(url);
-      const data = await response.json();
-      console.log(data);
-
-      if (data.Response === "False") {
-        setError({ show: true, msg: data.Error });
-        setLoading(false);
-      } else {
-        setMovie(data);
-        setLoading(false);
-      }
-    } catch (error) {
-      console.log(error);
-    }
-  };
-
-  useEffect(() => {
-    fetchMovie(`${API_ENDPOINT}&i=${id}`);
-  }, []);
+  const { loading, error, data: movie } = useFetch(`&i=${id}`);
 
   if (loading) {
     return <div className="loading"></div>;
