+++ 21-movie-db/setup/src/context.js
@@ -1,41 +1,15 @@
 import React, { useState, useContext, useEffect } from "react";
 // make sure to use https
-export const API_ENDPOINT = `https://www.omdbapi.com/?apikey=${process.env.REACT_APP_MOVIE_API_KEY}`;
-console.log(API_ENDPOINT);
+import useFetch from "./useFetch";
 
 const AppContext = React.createContext();
 
 const AppProvider = ({ children }) => {
-  const [loading, setLoading] = useState(true);
-  const [error, setError] = useState({ show: false, msg: "" });
-  const [movies, setMovies] = useState([]);
   const [query, setQuery] = useState("batman");
-
-  const fetchMovies = async (url) => {
-    setLoading(true);
-    try {
-      const response = await fetch(url);
-      const data = await response.json();
-      if (data.Response === "True") {
-        setMovies(data.Search);
-        setError({ show: false, msg: "" });
-      } else {
-        setError({ show: true, msg: "movie not found" });
-      }
-      setLoading(false);
-    } catch (error) {
-      console.log(error);
-    }
-  };
-
-  useEffect(() => {
-    fetchMovies(`${API_ENDPOINT}&s=${query}`);
-  }, [query]);
+  const { loading, error, data: movies } = useFetch(`&s=${query}`);
 
   return (
-    <AppContext.Provider
-      value={{ loading, error, movies, query, setQuery, setLoading, setError }}
-    >
+    <AppContext.Provider value={{ loading, error, movies, query, setQuery }}>
       {children}
     </AppContext.Provider>
   );
