+++ 21-movie-db/setup/src/useFetch.js
@@ -1,0 +1,34 @@
+import React, { useState, useEffect } from "react";
+export const API_ENDPOINT = `https://www.omdbapi.com/?apikey=${process.env.REACT_APP_MOVIE_API_KEY}`;
+
+const useFetch = (urlParams) => {
+  const [loading, setLoading] = useState(true);
+  const [error, setError] = useState({ show: false, msg: "" });
+  const [data, setData] = useState([]);
+
+  const fetchMovies = async (url) => {
+    setLoading(true);
+    try {
+      const response = await fetch(url);
+      const data = await response.json();
+      if (data.Response === "True") {
+        console.log(data);
+        setData(data.Search || data);
+        setError({ show: false, msg: "" });
+      } else {
+        setError({ show: true, msg: "movie not found" });
+      }
+      setLoading(false);
+    } catch (error) {
+      console.log(error);
+    }
+  };
+
+  useEffect(() => {
+    fetchMovies(`${API_ENDPOINT}${urlParams}`);
+  }, [urlParams]);
+
+  return { loading, error, data };
+};
+
+export default useFetch;
